+++ data_adapter/settings.py
@@ -2,6 +2,8 @@ import logging
 import os
 import pathlib
 
+import sqlalchemy as sa
+
 DEBUG = os.environ.get("DEBUG", "false") == "true"
 
 logger = logging.getLogger()
@@ -16,3 +18,7 @@ COLLECTIONS_DIR = (
 COLLECTION_JSON = "collection.json"
 
 DATABUS_ENDPOINT = "https://energy.databus.dbpedia.org/sparql"
+DATABASE_URL = os.environ.get("DATABASE_URL")
+
+if DATABASE_URL:
+    engine = sa.create_engine(DATABASE_URL)