+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Here is a template for new release sections
 -
 ```
 
+## [Unreleased]
+
+### Added
+- models for ES structure
+
 ## [0.4.1] - 2023-01-23
 
 ### Fixed
@@ -24,6 +29,9 @@ Here is a template for new release sections
 
 
 ## [0.4.0]
+
+### Added
+- checks for annotations of collection
 
 ### Changed
 - collection metadata holds data type (Note: Collections have to be reloaded)